Implement Roles Service, Project Service, & Refactor User Service

+++ internal/usecase/user_usecase.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 
 	"github.com/go-playground/validator/v10"
+	"github.com/sirupsen/logrus"
 
 	"github.com/tubagusmf/helpdesk-ticketing-nutech-integrasi-be/internal/helper"
 	"github.com/tubagusmf/helpdesk-ticketing-nutech-integrasi-be/internal/model"
@@ -23,7 +24,12 @@ func NewUserUsecase(userRepo model.IUserRepository) model.IUserUsecase {
 }
 
 func (u *UserUsecase) Login(ctx context.Context, in model.LoginInput) (string, error) {
+	log := logrus.WithFields(logrus.Fields{
+		"email": in.Email,
+	})
+
 	if err := validate.Struct(in); err != nil {
+		log.Error("Validation error", err)
 		return "", err
 	}
 
@@ -36,16 +42,26 @@ func (u *UserUsecase) Login(ctx context.Context, in model.LoginInput) (string, e
 		return "", errors.New("email or password is incorrect")
 	}
 
-	return helper.GenerateToken(*user)
+	token, err := helper.GenerateToken(*user)
+	if err != nil {
+		return "", err
+	}
+
+	return token, nil
 }
 
 func (u *UserUsecase) Create(ctx context.Context, in model.CreateUserInput) (string, error) {
+	log := logrus.WithFields(logrus.Fields{
+		"in": in,
+	})
+
 	if err := validate.Struct(in); err != nil {
 		return "", err
 	}
 
 	hashed, err := helper.HashRequestPassword(in.Password)
 	if err != nil {
+		log.Error(err)
 		return "", err
 	}
 
@@ -58,22 +74,53 @@ func (u *UserUsecase) Create(ctx context.Context, in model.CreateUserInput) (str
 	})
 
 	if err != nil {
+		log.Error(err)
 		return "", err
 	}
 
-	return helper.GenerateToken(*newUser)
+	accessToken, err := helper.GenerateToken(*newUser)
+	if err != nil {
+		return "", err
+	}
+
+	return accessToken, nil
 }
 
 func (u *UserUsecase) FindAll(ctx context.Context, filter model.User) ([]*model.User, error) {
-	return u.userRepo.FindAll(ctx, filter)
+	log := logrus.WithFields(logrus.Fields{
+		"filter": filter,
+	})
+
+	users, err := u.userRepo.FindAll(ctx, filter)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	return users, nil
 }
 
 func (u *UserUsecase) FindByID(ctx context.Context, id int64) (*model.User, error) {
-	return u.userRepo.FindByID(ctx, id)
+	log := logrus.WithFields(logrus.Fields{
+		"id": id,
+	})
+
+	user, err := u.userRepo.FindByID(ctx, id)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	return user, nil
 }
 
 func (u *UserUsecase) Update(ctx context.Context, id int64, in model.UpdateUserInput) error {
+	log := logrus.WithFields(logrus.Fields{
+		"id": id,
+	})
+
 	if err := validate.Struct(in); err != nil {
+		log.Error("Validation error", err)
 		return err
 	}
 
@@ -96,5 +143,19 @@ func (u *UserUsecase) Update(ctx context.Context, id int64, in model.UpdateUserI
 }
 
 func (u *UserUsecase) Delete(ctx context.Context, id int64) error {
+	log := logrus.WithFields(logrus.Fields{
+		"id": id,
+	})
+
+	user, err := u.userRepo.FindByID(ctx, id)
+	if err != nil {
+		log.Error("Failed to find user for deletion: ", err)
+		return err
+	}
+
+	if user == nil {
+		return errors.New("user not found")
+	}
+
 	return u.userRepo.Delete(ctx, id)
 }
